### 2025-02-21 #689 @rmgpanw
trud: Query the 'NHS TRUD API'
Submitting Author Name: Alasdair Warwick
Submitting Author Github Handle: @rmgpanw
Other Package Authors Github handles: (comma separated, delete if none) @abolvera, @cyu12, @rnluben
Repository:  https://github.com/rmgpanw/trud/
Submission type: Pre-submission

---

```
Package: trud
Title: Query the 'NHS TRUD API'
Version: 0.1.0.9000
    person(given = "Alasdair",
           family = "Warwick",
           email = "alasdair.warwick.19@ucl.ac.uk",
    person("Robert", "Luben", role = "aut",
    person("Abraham", "Olvera-Barrios", role = "aut",
    person("Chuin Ying", "Ung", role = "aut",
    )
Description: A convenient R interface to the 'National Health Service (NHS)
    Technology Reference Update Distribution (TRUD) API'. Retrieve available
    releases for items that you are subscribed to and download these with ease.
    For more information on the API, see <https://isd.digital.nhs.uk/trud/users/guest/filters/0/api>.
    cli,
    httr2,
    rvest,
    tidyselect
    https://github.com/rmgpanw/trud
    withr
```

`trud` wraps the [NHS TRUD API](https://isd.digital.nhs.uk/trud/users/guest/filters/0/api), which account holders to automate the download of technology reference files from NHS England.

The technology reference files from NHS England are widely used by researchers (both within and outside the UK) working with electronic health records data. The available items include various clinical coding terminologies such as ICD, Read, prescription codes and the SNOMED CT ontology (which is updated with new available releases every few months). While `trud` is a relatively small package, it provides a helpful aid to researchers working with these release items, greatly simplifying the process of automated updates to support downstream analyses.

There are no other similar R packages that I am aware of.

`trud` 0.1.0 is available on CRAN (accepted mid-2024)
