+++ NAMESPACE
@@ -1,8 +1,6 @@
 # Generated by roxygen2: do not edit by hand
 
-export("%>%")
 export(download_item)
 export(get_item_metadata)
 export(get_subscribed_metadata)
 export(trud_items)
-importFrom(magrittr,"%>%")
