+++ DESCRIPTION
@@ -21,12 +21,12 @@ Description: A convenient R interface to the 'National Health Service (NHS)
 License: MIT + file LICENSE
 Encoding: UTF-8
 Roxygen: list(markdown = TRUE)
-RoxygenNote: 7.2.3
+RoxygenNote: 7.3.2
+Depends: R (>= 4.2.0)
 Imports: 
     cli,
     dplyr,
     httr2,
-    magrittr,
     purrr,
     rlang,
     rvest,
