### 2025-04-27 66af0b3
update trud description in README and DESCRIPTION
