### 2025-04-27 8f5c914
add code of conduct, contributing and citation details
