### 2025-04-27 2b73198
update roxygen to use @returns instead of @return, and @examplesIf to run examples for pkgdown site only
