+++ R/get_item_metadata.R
@@ -7,20 +7,19 @@
 #' @param latest_only If `TRUE`, only metadata pertaining to the latest item
 #'   release will be retrieved. By default this is set to `FALSE`.
 #'
-#' @return A list.
+#' @returns A list.
 #' @export
 #'
-#' @examples
-#' \dontrun{
-#'  # Get metadata for Community Services Data Set pre-deadline extract XML Schema
-#'  get_item_metadata(394) |>
-#'    purrr::map_at("releases", \(release) purrr::map(release, names))
+#' @examplesIf identical(Sys.getenv("IN_PKGDOWN"), "true")
+#' # Get metadata for Community Services Data Set pre-deadline extract XML Schema
+#' get_item_metadata(394) |>
+#'   purrr::map_at("releases", \(release) purrr::map(release, names))
 #'
 #' # Include metadata for any previous releases using `latest_only = FALSE`
 #' get_item_metadata(394, latest_only = FALSE) |>
-#'    purrr::map_at("releases", \(release) purrr::map(release, names))
-#' }
+#'   purrr::map_at("releases", \(release) purrr::map(release, names))
 #'
+#' @examples
 #' # An informative error is raised if your API key is invalid or missing
 #' try(download_item(394, TRUD_API_KEY = "INVALID_API_KEY"))
 get_item_metadata <- function(item,
