### 2025-04-27 094dbfe
add getting started vignette and update README
