+++ README.md
@@ -16,16 +16,17 @@ status](https://www.r-pkg.org/badges/version/trud)](https://CRAN.R-project.org/p
 Status](https://www.repostatus.org/badges/latest/active.svg)](https://www.repostatus.org/#active)
 <!-- badges: end -->
 
-The goal of `trud` is to provide a convenient R interface to the [NHS
-TRUD API](https://isd.digital.nhs.uk/trud/users/guest/filters/0/api).
+The goal of `trud` is to provide a convenient R interface to the
+[National Health Service (NHS) England Technology Reference data Update
+Distribution
+(TRUD)](https://isd.digital.nhs.uk/trud/users/guest/filters/0/api).
 
-The National Health Service (NHS) England Technology Reference data
-Update Distribution (TRUD) service provides essential reference files
-that underpin a wide range of electronic health record (EHR) research,
-both in the UK and internationally. These files include clinical coding
-systems such as ICD, Read codes, prescription codes, and the SNOMED CT
-ontology, with regular updates to reflect new knowledge and changes in
-clinical practice. NHS TRUD content supports key areas like disease
+The NHS TRUD service provides essential reference files that underpin a
+wide range of electronic health record (EHR) areas, both in the UK and
+internationally. These files include clinical coding systems such as
+ICD, Read codes, prescription codes, and the SNOMED CT ontology, with
+regular updates to reflect new knowledge and changes in clinical
+practice. NHS TRUD content supports key research areas like disease
 phenotyping, cohort selection, epidemiology, health services research,
 and the development of risk prediction models.
 
@@ -54,126 +55,15 @@ Or you can install the development version of `trud` from
 pak::pak("rmgpanw/trud")
 ```
 
-You will also need to:
+You will also need to [sign up for a free
+account](https://isd.digital.nhs.uk/trud/users/guest/filters/0/account/form)
+with NHS TRUD
 
-- [Sign up for an
-  account](https://isd.digital.nhs.uk/trud/users/guest/filters/0/account/form)
-  with NHS TRUD
-- Set your API key (shown on your account information page) as an
-  environmental variable named `TRUD_API_KEY`. The best way to do this
-  is by placing your API key in a [`.Renviron`
-  file](https://rstats.wtf/r-startup.html#renviron).
+## Available functionality
 
-## Examples
-
-Retrieve available endpoints[^1]:
-
-``` r
-library(trud)
-
-trud_items()
-#> # A tibble: 73 × 2
-#>    item_number item_name                                                        
-#>          <int> <chr>                                                            
-#>  1         246 Cancer Outcomes and Services Data Set XML Schema                 
-#>  2         245 Commissioning Data Set XML Schema                                
-#>  3         599 Community Services Data Set Intermediate Database                
-#>  4         393 Community Services Data Set post-deadline extract XML Schema     
-#>  5         394 Community Services Data Set pre-deadline extract XML Schema      
-#>  6         391 Community Services Data Set XML Schema                           
-#>  7         248 Diagnostic Imaging Data Set XML Schema                           
-#>  8         239 dm+d XML transformation tool                                     
-#>  9        1859 Electronic Prescribing and Medicines Administration Data Sets XM…
-#> 10        1819 Emergency Care Data Set XML Schema                               
-#> # ℹ 63 more rows
-```
-
-Get metadata for an item:
-
-``` r
-get_item_metadata(394) |>
-  purrr::map_at("releases", \(release) purrr::map(release, names))
-#> $apiVersion
-#> [1] "1"
-#> 
[…]
