+++ inst/paper/paper.bib
@@ -1,0 +1,13 @@
+@manual{trud,
+  title={Query the NHS TRUD API},
+  author={Alasdair Warwick and Abraham Olvera-Barrios and Chuin Ying Ung and Robert Luben},
+  year={2024},
+  url={https://github.com/rmgpanw/trud}
+}
+
+@misc{trudwebsite,
+  title={NHS Technology Reference Data Update Distribution},
+  author={{NHS England}},
+  year={2024},
+  url={https://isd.digital.nhs.uk/trud}
+}
