### 2025-04-28 #689 @rmgpanw
Hi @maurolepore, I hope all is well. https://github.com/ropensci/software-review/issues/684 has now been approved - could this presubmission please be reconsidered?

Also please note that all pkgcheck checks are passing for me locally, however the pkgcheck GitHub Actions currently hangs (being investigated at https://github.com/ropensci-review-tools/pkgcheck/issues/235). Many thanks
