### 2025-05-05 #689 @maurolepore
@rmgpanw thanks for sharing your work with rOpenSci.

Here're my preliminary checks for discussion with the editorial board. If any of this helps as feedback, it's a good time to make changes.

### Preliminary checks:

    > Not grouped but the API is small so clear anyway.

---

#### Comments

* README has no examples.
* I see a non-standard style, e.g. in tests/testthat/test-get_item_metadata.R.
  Our guide recommends the tidyverse stryle guide and automating its
  implementation with the styler package or the air CLI.
* I see what seems to be a missplaced test -- an exported function that
  has its own test-file but is also tested in test-utils.R:
  https://github.com/rmgpanw/trud/blob/1565e8f7192418b035781de0376d1ec04797192a/tests/testthat/test-utils.R#L11C12-L11C29
* I see a call to `expect_error()` matching a regex that seems to be too
  specific here:
  https://github.com/rmgpanw/trud/blob/1565e8f7192418b035781de0376d1ec04797192a/tests/testthat/test-download_item.R#L7
  This can results in brittle tests. Other strategies:

    * Match a relevant piece only, e.g.:

    ```r
    # too specific
    expect_error(f(x), "This containts too many specific details")

    # match just a relevant piece
    expect_error(f(x), "too.*specific")
    ```

    * Use the `class` argument:

    ```r
    expect_error(f(x), class = "missing_api_key")
    ```

    * Use snapshots

    ```r
    expect_snapshot(error = TRUE, f(x))
    ```

---
