### 2025-05-05 64bb86e
relocate misplaced test for get_item_metadata()
