### 2025-05-05 eb46840
add example to README
