### 2025-05-05 32b5817
add error classes "missing_api_key" and "unrecognised_trud_item_release"
