+++ tests/testthat/test-utils.R
@@ -1,11 +1,14 @@
 test_that("get_trud_api_key() raises error with non-string argument", {
-  expect_error(get_trud_api_key(1),
-               "Argument `TRUD_API_KEY` must either be a string or `NULL`")
+  expect_error(
+    get_trud_api_key(1),
+    "Argument `TRUD_API_KEY` must either be a string or `NULL`"
+  )
 })
 
 test_that("get_trud_api_key() raises error with missing API key", {
   expect_error(get_trud_api_key(TRUD_API_KEY = "MISSING_API_KEY"),
-               class = "missing_api_key")
+    class = "missing_api_key"
+  )
 })
 
 test_that("validate_arg_item() raises error with non-integer argument", {
@@ -13,16 +16,17 @@ test_that("validate_arg_item() raises error with non-integer argument", {
 })
 
 
-test_that("validate_arg_directory() raises error with non-string argument",
-          {
-            expect_error(validate_arg_directory(1),
-                         "Argument `directory` must be a string")
+test_that("validate_arg_directory() raises error with non-string argument", {
+  expect_error(
+    validate_arg_directory(1),
+    "Argument `directory` must be a string"
+  )
 
-            expect_error(
-              validate_arg_directory(file.path(tempdir(), "invaliddirectory")),
-              "Argument `directory` must be a valid file path."
-            )
-          })
+  expect_error(
+    validate_arg_directory(file.path(tempdir(), "invaliddirectory")),
+    "Argument `directory` must be a valid file path."
+  )
+})
 
 test_that("trud_error_message() returns expected error messages", {
   response_without_json_body <- httr2::response(
@@ -33,8 +37,10 @@ test_that("trud_error_message() returns expected error messages", {
     body = raw()
   )
 
-  expect_match(trud_error_message(response_without_json_body)[2],
-               "Is the TRUD website down?")
+  expect_match(
+    trud_error_message(response_without_json_body)[2],
+    "Is the TRUD website down?"
+  )
 
   example_response <- httr2::response(
     status_code = 200,
@@ -44,29 +50,35 @@ test_that("trud_error_message() returns expected error messages", {
     body = raw()
   )
 
-  with_mocked_bindings({
-    status_400_result <- trud_error_message(example_response)
+  with_mocked_bindings(
+    {
+      status_400_result <- trud_error_message(example_response)
 
-    expect_match(status_400_result, "invalid API key")
-  }, try_resp_body_json = function(...) {
-    list(httpStatus = 400)
-  })
+      expect_match(status_400_result, "invalid API key")
+    },
+    try_resp_body_json = function(...) {
+      list(httpStatus = 400)
+    }
+  )
 
-  with_mocked_bindings({
-    status_404_result <- trud_error_message(example_response)
+  with_mocked_bindings(
+    {
+      status_404_result <- trud_error_message(example_response)
 
-    expect_match(
-      status_404_result[1],
-      "Either this item number does not exist, or you are not subscribed to it"
-    )
-  }, try_resp_body_json = function(...) {
-    list(httpStatus = 404)
-  })
-
+      expect_match(
+        status_404_result[1],
+        "Either this item number does not exist, or you are not subscribed to it"
+      )
+    },
+    try_resp_body_json = function(...) {
+      list(httpStatus = 404)
+    }
+  )
 })
 
-test_that("validate_arg_download_file() raises error with non-string argument",
-          {
-            expect_error(validate_arg_download_file(1),
-                         "Argument `download_file` must be a string.")
-          })
+test_that("validate_arg_download_file() raises error with non-string argument", {
+  expect_error(
+    validate_arg_download_file(1),
+    "Argument `download_file` must be a string."
+  )
+})
