+++ R/utils.R
@@ -1,8 +1,10 @@
 get_trud_api_key <- function(TRUD_API_KEY, call = rlang::caller_env()) {
   error_guidance_messages <-
-    c("i" = "Set your NHS TRUD API key as an environment variable using {.code Sys.setenv(TRUD_API_KEY='<<your-key>>')}, or preferably use a {.code .Renviron} file",
+    c(
+      "i" = "Set your NHS TRUD API key as an environment variable using {.code Sys.setenv(TRUD_API_KEY='<<your-key>>')}, or preferably use a {.code .Renviron} file",
       "i" = "To get an API key, first sign up for a NHS TRUD account at {.url https://isd.digital.nhs.uk/trud/users/guest/filters/0/account/form}",
-      "i" = "To find Your API key, log in and visit your account profile page ({.url https://isd.digital.nhs.uk/trud/users/authenticated/filters/0/account/manage}).")
+      "i" = "To find Your API key, log in and visit your account profile page ({.url https://isd.digital.nhs.uk/trud/users/authenticated/filters/0/account/manage})."
+    )
 
   if (is.null(TRUD_API_KEY)) {
     TRUD_API_KEY <- "TRUD_API_KEY"
@@ -20,10 +22,14 @@ get_trud_api_key <- function(TRUD_API_KEY, call = rlang::caller_env()) {
 
   TRUD_API_KEY <- Sys.getenv(TRUD_API_KEY)
   if (identical(TRUD_API_KEY, "")) {
-    cli::cli_abort(message = c("x" = "Can't find NHS TRUD API key",
-                               error_guidance_messages),
-                   class = "missing_api_key",
-                   call = call)
+    cli::cli_abort(
+      message = c(
+        "x" = "Can't find NHS TRUD API key",
+        error_guidance_messages
+      ),
+      class = "missing_api_key",
+      call = call
+    )
   }
 
   invisible(TRUD_API_KEY)
@@ -36,8 +42,7 @@ trud_error_message <- function(resp) {
     return(resp)
   }
 
-  switch(
-    as.character(resp$httpStatus),
+  switch(as.character(resp$httpStatus),
     "404" = c(
       "x" = stringr::str_replace(resp$message, "API key .*,", "supplied API key,"),
       "i" = "Either this item number does not exist, or you are not subscribed to it",
@@ -68,9 +73,13 @@ try_resp_body_json <- function(resp) {
 
 validate_arg_item <- function(item, call = rlang::caller_env()) {
   if (!rlang::is_scalar_integerish(item)) {
-    cli::cli_abort(c("x" = "Argument {.code item} must be an integer.",
-                     "i" = "Use {.code trud_items()} to see available items."),
-                   call = call)
+    cli::cli_abort(
+      c(
+        "x" = "Argument {.code item} must be an integer.",
+        "i" = "Use {.code trud_items()} to see available items."
+      ),
+      call = call
+    )
   }
 }
 
@@ -85,9 +94,13 @@ validate_arg_directory <- function(directory, call = rlang::caller_env()) {
     )
 
   if (!rlang::is_string(directory)) {
-    cli::cli_abort(c("x" = "Argument {.code directory} must be a string.",
-                     directory_arg_example),
-                   call = call)
+    cli::cli_abort(
+      c(
+        "x" = "Argument {.code directory} must be a string.",
+        directory_arg_example
+      ),
+      call = call
+    )
   }
 
   if (!dir.exists(directory)) {
@@ -106,20 +119,23 @@ validate_arg_download_file <- function(download_file, call = rlang::caller_env()
   valid_download_file_values <- c("archive", "checksum", "signature", "publicKey")
 
   if (!rlang::is_string(download_file)) {
-    cli::cli_abort(c(
[…]
