### 2025-05-05 #689 @rmgpanw
Thanks @maurolepore! I have now:

- Added a README example ([eb46840](https://github.com/rmgpanw/trud/commit/eb4684060e2e25a2859a0a90fe095a5a0ef010ba))
- Relocated the misplaced test ([64bb86e](https://github.com/rmgpanw/trud/commit/64bb86ed5f9440f3c1cc09a1a2b3a032ec11da39))
- Amended the overly `expect_error()` message to use the `class` argument ([32b5817](https://github.com/rmgpanw/trud/commit/32b58177bd3c5847f5e9a31e3e3d8eecccab9f53))
- Applied the styler package ([c4a8a27](https://github.com/rmgpanw/trud/commit/c4a8a27839947a7ace873a09a1647a2cf6a34ffd))

Regarding the pkgcheck issues, trud require an API key which I have saved as a secret in the repo (with all tests passing). Could you please advise how I should address this so that the tests run as expected in the rOpenSci environment?

Also please note that all pkgcheck checks are passing for me locally, however the pkgcheck GitHub Actions currently hangs (being investigated at https://github.com/ropensci-review-tools/pkgcheck/issues/235). Many thanks
