### 2025-05-06 #689 @maurolepore
@rmgpanw  thanks so much for following up promptly.

RE where to find wisdome about how to handle API keys in challenging CI environments:

* https://books.ropensci.org/http-testing/
* https://httr2.r-lib.org/articles/wrapping-apis.html

It's also a good idea to see how other packages similar to yours do it. Some API keys are  particularly sensitive because you need to pay for them, or expose sentitive data, or a million other reasons. But some other API keys anyone can get easily for free. Those ones are typically not so sensitive and maybe there is some way to work with them in a less rigid way? Not sure so after doing your research I would post specific questions to @maelle or @mpadge.

Given my EiC rotation ended, and considering you have made great progress, I suggest we close this issue so it's not hanging indefinitely.

When you're ready please open an issue for a full submission and reference this pre-submission for the consideration of the EiC.

It's been great working with you. Best luck!!
